String to Long

+++ src/main/kotlin/no/nav/hjelpemidler/brille/joarkref/JoarkrefRiver.kt
@@ -52,7 +52,7 @@ class JoarkrefRiver(
         runBlocking {
             withContext(Dispatchers.IO) {
                 transaction(databaseContext) { ctx ->
-                    ctx.joarkrefStore.lagreJoarkRef(packet.sakId, packet.joarkRef)
+                    ctx.joarkrefStore.lagreJoarkRef(packet.sakId.toLong(), packet.joarkRef.toLong())
                 }
             }
         }
